// webapp/agnor-backend/src/citas/dto/cita.dto.ts
export class CitaDto{
    id: number;
    idcontacto: number;
    titulo: string;
    descripcion: string;
    fecha: string;
    estado: string;
    grupo: number;
    fecha_registro: string;
    seguimiento: string;
    elaborado: string;
    tipodecita: string;
    apoyos: string;
    direccioncit: string;
}