// webapp/agnor-backend/src/citas/interface/cita.interface.ts
export interface CitaI {
    id?: number;
    idcontacto?: number;
    titulo?: string;
    descripcion?: string;
    fecha?: string;
    estado?: string;
    grupo?: number;
    fecha_registro?: string;
    seguimiento?: string;
    elaborado?: string;
    tipodecita?: string;
    apoyos?: string;
    direccioncit?: string;
}